bg/js/popup.js文件拆分为文件夹
