+++ src/bg/js/popup/renderMomentCircleHtml.js
@@ -1,0 +1,6 @@
+export default  function renderMomentCircleHtml() {
+    chrome.storage.local.get(["AcMomentCircle1"], function (data) {
+      $("#pop-push-momentcircle").append(data.AcMomentCircle1);
+    });
+  }
+
