+++ src/bg/popup.html
@@ -154,7 +154,7 @@
     <script src="../lib/jquery-3.5.0.min.js"></script>
     <script src="js/lib/purify.min.js"></script>
     <script src="../common/utils.js"></script>
-    <script src="js/popup.js"></script>
+    <script type='module' src="js/popup/index.js"></script>
     <script src="js/lib/md5.js"></script>
     <script type="text/javascript" src="js/jquery.mloading.js"></script>
     <script type="text/javascript" src="../mdui/js/mdui.min.js"></script>
